testing using categories to separate two blog streams
the site needs to have a blog but also has a case studies page. Both 
those pages would be nice in the jekyll blog format, allowing users to 
simply add posts to a _posts directory in markdown to allow them to 
update it. The jekyll template can then nicely line them up by date and 
preview them for people to browse nicely.